Make broker compatible with the new core
