+++ rumq-broker/src/connection.rs
@@ -1,5 +1,5 @@
 use derive_more::From;
-use rumq_core::{connack, Packet, Connect, ConnectReturnCode};
+use rumq_core::mqtt4::{connack, Packet, Connect, ConnectReturnCode};
 use tokio::sync::mpsc::{channel, Receiver, Sender};
 use tokio::sync::mpsc::error::SendError;
 use tokio::stream::iter;
