+++ rumq-broker/src/router.rs
@@ -1,5 +1,5 @@
 use derive_more::From;
-use rumq_core::{has_wildcards, matches, QoS, Packet, Connect, Publish, Subscribe, Unsubscribe};
+use rumq_core::mqtt4::{has_wildcards, matches, publish, QoS, Packet, Connect, Publish, Subscribe, Unsubscribe};
 use tokio::sync::mpsc::{Receiver, Sender};
 use tokio::sync::mpsc::error::TrySendError;
 use tokio::select;
@@ -27,9 +27,9 @@ pub enum RouterMessage {
     /// Client id and connection handle
     Connect(Connection),
     /// Packet
-    Packet(rumq_core::Packet),
+    Packet(Packet),
     /// Packets
-    Packets(VecDeque<rumq_core::Packet>),
+    Packets(VecDeque<Packet>),
     /// Disconnects a client from active connections list. Will handling
     Death(String),
     /// Pending messages of the previous connection
@@ -37,12 +37,12 @@ pub enum RouterMessage {
 }
 
 pub struct Connection {
-    pub connect: rumq_core::Connect,
+    pub connect: Connect,
     pub handle: Option<Sender<RouterMessage>>
 }
 
 impl Connection {
-    pub fn new(connect: rumq_core::Connect, handle: Sender<RouterMessage>) -> Connection {
+    pub fn new(connect: Connect, handle: Sender<RouterMessage>) -> Connection {
         Connection {
             connect,
             handle: Some(handle)
@@ -59,7 +59,7 @@ impl fmt::Debug for Connection {
 #[derive(Debug)]
 struct ActiveConnection {
     pub state: MqttState,
-    pub outgoing: VecDeque<rumq_core::Packet>,
+    pub outgoing: VecDeque<Packet>,
     tx: Sender<RouterMessage>
 }
 
@@ -153,7 +153,7 @@ impl Router {
                         error!("Routing error = {:?}", e);
                     }
                 }
-                o = interval.next() => {
+                _ = interval.next() => {
                     for (_, connection) in self.active_connections.iter_mut() {
                         let pending = connection.outgoing.split_off(0);
                         if pending.len() > 0 {
@@ -473,7 +473,7 @@ impl Router {
                 let message = mem::replace(&mut will.message, "".to_owned());
                 let qos = will.qos;
 
-                let publish = rumq_core::publish(topic, qos, message);
+                let publish = publish(topic, qos, message);
                 self.match_subscriptions(&id, publish);
             }
